Added dependencies, moved scraper files to new folder, formatted spacing in views.py
