+++ requirements.txt
@@ -12,3 +12,6 @@ requests==2.18.4
 six==1.11.0
 urllib3==1.22
 whitenoise==3.3.1
+beautifulsoup4==4.6.0
+bs4==0.0.1
+lxml==4.2.1
