+++ tigerpath/scraper/scrape_all.py
