+++ tigerpath/scraper/scrape_validate.py
